Move replace cue color dialog from main menu bar to preferences
