+++ src/preferences/dialog/dlgprefcolorsdlg.ui
@@ -26,6 +26,30 @@
       <string>Colors</string>
      </property>
      <layout class="QGridLayout" name="gridLayout">
+      <item row="1" column="2">
+       <widget class="QPushButton" name="pushButtonEditHotcuePalette">
+        <property name="text">
+         <string>Edit…</string>
+        </property>
+       </widget>
+      </item>
+      <item row="3" column="1">
+       <widget class="QComboBox" name="comboBoxHotcueDefaultColor"/>
+      </item>
+      <item row="0" column="2">
+       <widget class="QPushButton" name="pushButtonEditTrackPalette">
+        <property name="text">
+         <string>Edit…</string>
+        </property>
+       </widget>
+      </item>
+      <item row="0" column="0">
+       <widget class="QLabel" name="labelTrackColors">
+        <property name="text">
+         <string>Track palette</string>
+        </property>
+       </widget>
+      </item>
       <item row="1" column="1">
        <widget class="QComboBox" name="comboBoxHotcueColors">
         <property name="sizePolicy">
@@ -43,24 +67,10 @@
         </property>
        </widget>
       </item>
-      <item row="0" column="0">
-       <widget class="QLabel" name="labelTrackColors">
-        <property name="text">
-         <string>Track palette</string>
-        </property>
-       </widget>
-      </item>
-      <item row="2" column="0">
+      <item row="3" column="0">
        <widget class="QLabel" name="labelHotcueDefaultColor">
         <property name="text">
          <string>Hotcue default color</string>
-        </property>
-       </widget>
-      </item>
-      <item row="1" column="2">
-       <widget class="QPushButton" name="pushButtonEditHotcuePalette">
-        <property name="text">
-         <string>Edit…</string>
         </property>
        </widget>
       </item>
@@ -74,13 +84,10 @@
         </property>
        </widget>
       </item>
-      <item row="2" column="1" colspan="2">
-       <widget class="QComboBox" name="comboBoxHotcueDefaultColor"/>
-      </item>
-      <item row="0" column="2">
-       <widget class="QPushButton" name="pushButtonEditTrackPalette">
+      <item row="3" column="2">
+       <widget class="QPushButton" name="pushButtonReplaceCueColor">
         <property name="text">
-         <string>Edit…</string>
+         <string>Replace...</string>
         </property>
        </widget>
       </item>
